requirements.txt
google-genai
pandas
openpyxl
tqdm
streamlit
plotly